refactor to split otu search and status run to reduce api hit
